+++ pkg/config/dynamicloader.go
@@ -9,6 +9,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/grafana/agent/pkg/config/instrumentation"
 	v2 "github.com/grafana/agent/pkg/integrations/v2"
 	"github.com/grafana/agent/pkg/logs"
 	"github.com/grafana/agent/pkg/metrics"
@@ -106,6 +107,8 @@ func (c *DynamicLoader) LoadConfigByPath(path string) error {
 	default:
 		return fmt.Errorf("config path must start with file:// or s3://, not %s", path)
 	}
+
+	instrumentation.ConfigMetrics.InstrumentConfig(buf)
 
 	cl := &LoaderConfig{}
 	err = yaml.Unmarshal(buf, cl)