+++ CHANGELOG.md
@@ -67,6 +67,8 @@ v0.28.0 (2022-09-29)
 
 - Operator no longer set the `SecurityContext.Privileged` flag in the `config-reloader` container. (@hsyed-dojo)
 
+- Add metrics for config reloads and config hash (@jcreixell)
+
 v0.27.1 (2022-09-09)
 --------------------